extract look command
