+++ errors/roomErrors.js
